+++ bcipy/helpers/visualization.py
@@ -210,6 +210,7 @@ def visualize_gaze(
             colorbar=True)
 
     if raw_plot:
+        # ax.scatter(lx, range(len(lx)), c='r', s=1)
         ax.scatter(lx, ly, c='r', s=1)
         ax.scatter(rx, ry, c='b', s=1)
 
@@ -282,10 +283,13 @@ def visualize_gaze_inquiries(
     ly = 1 - ly
     ry = 1 - ry
 
-    if means is not None:
-        means[:, 0] = np.clip(means[:, 0], 0, 1)
-        means[:, 1] = np.clip(means[:, 1], 0, 1)
-        means[:, 1] = 1 - means[:, 1]
+    # Define mns as a copy of means to avoid modifying the original array
+    mns = np.copy(means)
+
+    if mns is not None:
+        mns[:, 0] = np.clip(mns[:, 0], 0, 1)
+        mns[:, 1] = np.clip(mns[:, 1], 0, 1)
+        mns[:, 1] = 1 - mns[:, 1]
 
     # scale the eye data to the image
     fig, ax = plt.subplots()
@@ -310,11 +314,93 @@ def visualize_gaze_inquiries(
             colorbar=True)
 
     if raw_plot:
-        # ax.scatter(lx, ly, c='lightcoral', s=1)
+        ax.scatter(lx, ly, c='lightcoral', s=1)
         ax.scatter(rx, ry, c='bisque', s=1)
 
-    if means is not None:
-        for i, (mean, cov) in enumerate(zip(means, covs)):
+    if mns is not None:
+        for i, (mean, cov) in enumerate(zip(mns, covs)):
+            v, w = linalg.eigh(cov)
+            v = 2.0 * np.sqrt(2.0) * np.sqrt(v)
+            u = w[0] / linalg.norm(w[0])
+
+            # Plot an ellipse to show the Gaussian component
+            angle = np.arctan(u[1] / u[0])
+            angle = 180.0 * angle / np.pi  # convert to degrees
+            ell = Ellipse(mean, v[0], v[1], angle=180.0 + angle, color='navy')
+            ell.set_clip_box(ax)
+            ell.set_alpha(0.5)
+            ax.add_artist(ell)
+
+    ax.set_xticks([])
+    ax.set_yticks([])
+    ax.set_xticklabels([])
+    ax.set_yticklabels([])
+
+    plt.title(f'{title}Plot')
+
+    if save_path is not None:
+        plt.savefig(f"{save_path}/{title.lower().replace(' ', '_')}plot.png", dpi=fig.dpi)
+
+    if show:
+        plt.show()
+    else:
+        plt.close()
+
+    return fig
+
+
+def visualize_pupil_size(
+        means: Optional[np.ndarray] = None,
+        covs: Optional[np.ndarray] = None,
+        save_path: Optional[str] = None,
+        show: Optional[bool] = False,
+        img_path: Optional[str] = None,
+        screen_size: Tuple[int, int] = (1920, 1080),
+        heatmap: Optional[bool] = False,
+        raw_plot: Optional[bool] = False) -> Figure:
+    """Visualize Gaze Inquiries.
+
+    Assumes that the data is collected using BciPy and a Tobii-nano eye tracker. The default
+    image used is for the matrix calibration task on a 1920x1080 screen.
+
+    Generates a comparative matrix figure following the execution of offline analysis. Given a set of
+    trailed data (left & right eye), the gaze distribution for each prompted symbol are plotted, along
+    with the contour plots of mean and covariances calculated by the Gaussian Mixture Model.
+    The figures may be saved or shown in a window.
+
+    Returns a list of the figure handles created.
+
+    Parameters
+    ----------
+    means: Optional[np.ndarray]: Means of the Gaussian Mixture Model
+    covs: Optional[np.ndarray]: Covariances of the Gaussian Mixture Model
+    save_path: Optional[str]: optional path to a save location of the figure generated
+    show: Optional[bool]: whether or not to show the figures generated. Default: False
+    img_path: Optional[str]: Image to be used as the background. Default: matrix.png
+    screen_size: Optional[Tuple[int, int]]: Size of the screen used for Calibration/Copy
+        Phrase tasks.
+        Default: (1920, 1080)
+    heatmap: Optional[bool]: Whether or not to plot the heatmap. Default: False
+    raw_plot: Optional[bool]: Whether or not to plot the raw gaze data. Default: False
+    """
+
+    title = 'Pupil Size '
+    img = plt.imread(img_path)
+
+    # Define mns as a copy of means to avoid modifying the original array
+    mns = np.copy(means)
+
+    if mns is not None:
+        # mns[:, 0] = np.clip(mns[:, 0], 0, 1)
+        # mns[:, 1] = np.clip(mns[:, 1], 0, 1)
+        mns[:, 1] = 1 - mns[:, 1]
+
+    # scale the eye data to the image
+    fig, ax = plt.subplots()
+    ax.imshow(img, extent=[0, 1, 0, 1])
+
+    if mns is not None:
+        for i, (mean, cov) in enumerate(zip(mns, covs)):
             v, w = linalg.eigh(cov)
             v = 2.0 * np.sqrt(2.0) * np.sqrt(v)
             u = w[0] / linalg.norm(w[0])
@@ -346,8 +432,7 @@ def visualize_gaze_inquiries(
 
 
 def visualize_centralized_data(
-        left_eye: np.ndarray,
-        right_eye: np.ndarray,
+        gaze_data: np.ndarray,
         save_path: Optional[str] = None,
         show: Optional[bool] = False,
         img_path: Optional[str] = None,
@@ -369,8 +454,7 @@ def visualize_centralized_data(
 
     Parameters
     ----------
-    left_eye: (np.ndarray): Data array for the centralized left eye data.
-    right_eye: (np.ndarray): Data array for the centralized right eye data.
+    gaze_data: (np.ndarray): Data array for the centralized left eye data.
     save_path: Optional[str]: optional path to a save location of the figure generated
     show: Optional[bool]: whether or not to show the figures generated. Default: False
     img_path: Optional[str]: Image to be used as the background. Default: matrix.png
@@ -390,33 +474,12 @@ def visualize_centralized_data(
 
     # Transform the eye data to fit the display. Clip values > 1 and < -1
     # The idea here is to have the center at (0,0)
-    lx = np.clip(left_eye[:, 0], -1, 1)
-    ly = np.clip(left_eye[:, 1], -1, 1)
-    rx = np.clip(right_eye[:, 0], -1, 1)
-    ry = np.clip(right_eye[:, 1], -1, 1)
+    dx = np.clip(gaze_data[:, 0], -1, 1)
+    dy = np.clip(gaze_data[:, 1], -1, 1)
     ax.imshow(img, extent=[-1, 1, -1, 1])
 
-    if heatmap:
-        # create a dataframe making a column for each x, y pair for both eyes and a column for the eye (left or right)
-        df = pd.DataFrame({
-            'x': np.concatenate((lx, rx)),
-            'y': np.concatenate((ly, ry)),
-            'eye': ['left'] * len(lx) + ['right'] * len(rx)
-        })
-        ax = sns.kdeplot(
-            data=df,
-            hue='eye',
-            x='x',
-            y='y',
-            fill=False,
-            thresh=0.05,
-            levels=10,
-            cmap="mako",
-            colorbar=True)
-
     if raw_plot:
-        ax.scatter(lx, ly, c='r', s=1)
-        ax.scatter(rx, ry, c='b', s=1)
+        ax.scatter(dx, dy, c='b', s=1)
 
     ax.set_xticks([])
     ax.set_yticks([])
@@ -459,10 +522,10 @@ def visualize_results_all_symbols(
 
     Parameters
     ----------
-    left_eye: (np.ndarray): Data array for the left eye data.
-    right_eye: (np.ndarray): Data array for the right eye data.
-    means: Optional[np.ndarray]: Means of the Gaussian Mixture Model
-    covs: Optional[np.ndarray]: Covariances of the Gaussian Mixture Model
+    left_eye_all: (np.ndarray): Data array for the left eye data, for all symbols.
+    right_eye_all: (np.ndarray): Data array for the right eye data, for all symbols.
+    means_all: Optional[np.ndarray]: Means of the Gaussian Mixture Model, for all symbols
+    covs_all: Optional[np.ndarray]: Covariances of the Gaussian Mixture Model, for all symbols
     save_path: Optional[str]: optional path to a save location of the figure generated
     show: Optional[bool]: whether or not to show the figures generated. Default: False
     img_path: Optional[str]: Image to be used as the background. Default: matrix.png
@@ -492,10 +555,13 @@ def visualize_results_all_symbols(
         ly = 1 - ly
         ry = 1 - ry
 
-        if means is not None:
-            means[:, 0] = np.clip(means[:, 0], 0, 1)
-            means[:, 1] = np.clip(means[:, 1], 0, 1)
-            means[:, 1] = 1 - means[:, 1]
+        # Define mns as a copy of means to avoid modifying the original array
+        mns = np.copy(means)
+
+        if mns is not None:
+            mns[:, 0] = np.clip(mns[:, 0], 0, 1)
+            mns[:, 1] = np.clip(mns[:, 1], 0, 1)
+            mns[:, 1] = 1 - mns[:, 1]
 
         if heatmap:
             # create a dataframe making a column for each x, y pair for both eyes and
@@ -517,11 +583,11 @@ def visualize_results_all_symbols(
                 colorbar=True)
 
         if raw_plot:
-            # ax.scatter(lx, ly, c='lightcoral', s=1)
+            ax.scatter(lx, ly, c='lightcoral', s=1)
             ax.scatter(rx, ry, c='bisque', s=1)
 
-        if means is not None:
-            for i, (mean, cov) in enumerate(zip(means, covs)):
+        if mns is not None:
+            for i, (mean, cov) in enumerate(zip(mns, covs)):
                 v, w = linalg.eigh(cov)
                 v = 2.0 * np.sqrt(2.0) * np.sqrt(v)
                 u = w[0] / linalg.norm(w[0])
@@ -751,12 +817,21 @@ def visualize_gaze_accuracies(accuracy_dict: Dict[str, np.ndarray],
 
     Returns a list of the figure handles created.
     """
+    title = 'Overall Accuracy: '
 
     fig, ax = plt.subplots()
     ax.bar(accuracy_dict.keys(), accuracy_dict.values())
     ax.set_xlabel('Symbol')
     ax.set_ylabel('Accuracy')
-    ax.set_title('Overall Accuracy: ' + str(round(accuracy, 2)))
+    ax.set_title(title + str(round(accuracy, 2)))
+
+    if save_path is not None:
+        plt.savefig(f"{save_path}/{title.lower().replace(' ', '_').replace(':', '')}plot.png", dpi=fig.dpi)
+
+    if show:
+        plt.show()
+    else:
+        plt.close()
 
     return fig