+++ bcipy/signal/model/pca_rda_kde/pca_rda_kde.py
@@ -114,7 +114,7 @@ class PcaRdaKdeModel(SignalModel):
         auc = -tmp
         return ModelEvaluationReport(auc)
 
-    def predict(self, data: np.array, inquiry: List[str], symbol_set: List[str]) -> np.array:
+    def compute_likelihood_ratio(self, data: np.array, inquiry: List[str], symbol_set: List[str]) -> np.array:
         """
         For each trial in `data`, compute a likelihood ratio to update that symbol's probability.
         Rather than just computing an update p(e|l=+) for the seen symbol and p(e|l=-) for all unseen symbols,
@@ -146,9 +146,9 @@ class PcaRdaKdeModel(SignalModel):
         likelihood_ratios = np.ones(len(symbol_set))
         for idx in range(len(subset_likelihood_ratios)):
             likelihood_ratios[symbol_set.index(inquiry[idx])] *= subset_likelihood_ratios[idx]
-        return likelihood_ratios
+        return likelihood_ratios   # used in multimodal update
 
-    def predict_proba(self, data: np.ndarray) -> np.ndarray:
+    def compute_class_probabilities(self, data: np.ndarray) -> np.ndarray:
         """Converts log likelihoods from model into class probabilities.
 
         Returns:
@@ -169,8 +169,49 @@ class PcaRdaKdeModel(SignalModel):
         denom = np.logaddexp(log_post_0, log_post_1)
         log_post_0 -= denom
         log_post_1 -= denom
-        posterior = np.exp(np.stack([log_post_0, log_post_1], axis=-1))
-        return posterior
+        log_posterior = np.stack([log_post_0, log_post_1], axis=-1)
+        return log_posterior
+
+    def evaluate_likelihood(self, data: np.ndarray) -> np.ndarray:
+        """
+        Calculates log(p(e | l)) for each trial in the data.
+        p(e | l=1), p(e | l=0)
+        """
+        if not self.ready_to_predict:
+            raise SignalException("must use model.fit() before model.predict_proba()")
+
+        log_scores_class_0 = self.model.transform(data)[:, 0]
+        log_scores_class_1 = self.model.transform(data)[:, 1]
+        return np.stack([log_scores_class_0, log_scores_class_1], axis=-1)
+
+    def predict(self, data: np.ndarray) -> np.ndarray:
+        """Predict the most likely label for each trial in the data.
+
+        Returns:
+            predictions (np.ndarray): shape (num_items,) - the predicted label for each item.
+        """
+        if not self.ready_to_predict:
+            raise SignalException("must use model.fit() before model.predict()")
+
+        posterior = self.compute_class_probabilities(data)
+        predictions = np.argmax(posterior, axis=1)
+        return predictions
+
+    def predict_proba(self, data: np.ndarray) -> np.ndarray:
+        """Converts log likelihoods from model into class probabilities.
+
+        Returns:
+            posterior (np.ndarray): shape (num_items, 2) - for each item, the model's predicted
+                probability for the two labels.
+        """
+        if not self.ready_to_predict:
+            raise SignalException("must use model.fit() before model.predict_proba()")
+
+        # Model originally produces p(eeg | label). We want p(label | eeg):
+        #
+        # p(l=1 | e) = p(e | l=1) p(l=1) / p(e)
+        # log(p(l=1 | e)) = log(p(e | l=1)) + log(p(l=1)) - log(p(e))
+        return self.compute_class_probabilities(data)
 
     def save(self, path: Path) -> None:
         """Save model weights (e.g. after training) to `path`"""